Extract Shapes module
